# whatsapp-api/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ClientStatus": {
				"type": "string",
				"enum": [
					"initializing",
					"qr_ready",
					"authenticated",
					"ready",
					"disconnected",
					"auth_failure"
				]
			},
			"StatusResponse": {
				"description": "GET /status response",
				"properties": {
					"status": {
						"$ref": "#/components/schemas/ClientStatus",
						"description": "Current lifecycle state of the WhatsApp client"
					},
					"qrAvailable": {
						"type": "boolean",
						"description": "Whether a QR code PNG is available at GET /qr"
					},
					"ready": {
						"type": "boolean",
						"description": "Convenience flag: true only when status === \"ready\""
					}
				},
				"required": [
					"status",
					"qrAvailable",
					"ready"
				],
				"type": "object",
				"additionalProperties": false
			},
			"QrJsonResponse": {
				"description": "GET /qr?format=json response",
				"properties": {
					"qr": {
						"type": "string",
						"description": "Base64 data-URL, e.g. \"data:image/png;base64,...\""
					}
				},
				"required": [
					"qr"
				],
				"type": "object",
				"additionalProperties": false
			},
			"QrUnavailableError": {
				"properties": {
					"error": {
						"type": "string"
					},
					"retryAfterSeconds": {
						"type": "number",
						"format": "double",
						"description": "Seconds until the client is expected to be ready (503 responses only)"
					},
					"hint": {
						"type": "string"
					}
				},
				"required": [
					"error",
					"hint"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponse": {
				"properties": {
					"error": {
						"type": "string"
					},
					"retryAfterSeconds": {
						"type": "number",
						"format": "double",
						"description": "Seconds until the client is expected to be ready (503 responses only)"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ServiceUnavailableError": {
				"description": "503 Service Unavailable — WhatsApp client is not in the 'ready' state.\nShared across controllers so tsoa sees exactly one model definition.",
				"properties": {
					"error": {
						"type": "string"
					},
					"retryAfterSeconds": {
						"type": "number",
						"format": "double",
						"description": "Seconds until the client is expected to be ready (503 responses only)"
					}
				},
				"required": [
					"error",
					"retryAfterSeconds"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LastMessage": {
				"properties": {
					"body": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double",
						"description": "Unix epoch seconds"
					},
					"fromMe": {
						"type": "boolean"
					}
				},
				"required": [
					"body",
					"type",
					"timestamp",
					"fromMe"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChatItem": {
				"properties": {
					"id": {
						"type": "string",
						"description": "WhatsApp chat ID, e.g. \"16073041892@c.us\" or \"123456789@g.us\""
					},
					"name": {
						"type": "string"
					},
					"isGroup": {
						"type": "boolean"
					},
					"unreadCount": {
						"type": "number",
						"format": "double"
					},
					"timestamp": {
						"type": "number",
						"format": "double",
						"description": "Unix epoch seconds of last activity"
					},
					"lastMessage": {
						"allOf": [
							{
								"$ref": "#/components/schemas/LastMessage"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"id",
					"name",
					"isGroup",
					"unreadCount",
					"timestamp",
					"lastMessage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChatsResponse": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"chats": {
						"items": {
							"$ref": "#/components/schemas/ChatItem"
						},
						"type": "array"
					},
					"hasMore": {
						"type": "boolean",
						"description": "True when more chats exist beyond the current page"
					}
				},
				"required": [
					"count",
					"chats",
					"hasMore"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MessageItem": {
				"properties": {
					"id": {
						"type": "string",
						"description": "Serialised message ID"
					},
					"from": {
						"type": "string"
					},
					"to": {
						"type": "string"
					},
					"body": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double",
						"description": "Unix epoch seconds"
					},
					"fromMe": {
						"type": "boolean"
					},
					"hasMedia": {
						"type": "boolean"
					},
					"author": {
						"type": "string",
						"nullable": true,
						"description": "Present for group messages; null otherwise"
					},
					"mimeType": {
						"type": "string",
						"description": "MIME type of attached media — only present when hasMedia is true"
					},
					"filename": {
						"type": "string",
						"description": "Original filename of attached media — only present when hasMedia is true"
					}
				},
				"required": [
					"id",
					"from",
					"to",
					"body",
					"type",
					"timestamp",
					"fromMe",
					"hasMedia",
					"author"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MessagesResponse": {
				"properties": {
					"chatId": {
						"type": "string"
					},
					"count": {
						"type": "number",
						"format": "double"
					},
					"messages": {
						"items": {
							"$ref": "#/components/schemas/MessageItem"
						},
						"type": "array"
					},
					"hasMore": {
						"type": "boolean",
						"description": "True when more messages exist beyond the current page"
					}
				},
				"required": [
					"chatId",
					"count",
					"messages",
					"hasMore"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "X-Api-Token",
				"description": "API_TOKEN from .env — set in X-Api-Token request header"
			},
			"basicAuth": {
				"type": "http",
				"scheme": "basic",
				"description": "BASIC_AUTH_USERNAME / BASIC_AUTH_PASSWORD from .env — standard HTTP Basic Auth"
			}
		}
	},
	"info": {
		"title": "WhatsApp Neo API",
		"version": "2.0.0",
		"description": "REST API for WhatsApp via whatsapp-web.js + remote Chromium CDP.\n\nAll routes except `/health` require either a Bearer token or Basic Auth:\n```\nX-Api-Token: <API_TOKEN>\nAuthorization: Basic <base64>\n```",
		"contact": {}
	},
	"paths": {
		"/status": {
			"get": {
				"operationId": "FetchStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StatusResponse"
								}
							}
						}
					}
				},
				"description": "Get the current WhatsApp client connection status.\n\nPossible states:\n- `initializing` — client is starting up / connecting to Chromium\n- `qr_ready` — QR code is available at GET /qr, waiting for phone scan\n- `authenticated` — credentials accepted, WhatsApp Web stores loading\n- `ready` — fully operational, API accepts all requests\n- `disconnected` — connection lost, reconnect scheduled\n- `auth_failure` — authentication error, manual intervention needed",
				"summary": "Get WhatsApp client status",
				"tags": [
					"System"
				],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"basicAuth": []
					}
				],
				"parameters": []
			}
		},
		"/qr": {
			"get": {
				"operationId": "GetQr",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QrJsonResponse"
								}
							}
						}
					},
					"409": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QrUnavailableError"
								}
							}
						}
					}
				},
				"description": "Get the current WhatsApp QR code.\n\nReturns `409` when no QR is available (already authenticated, or client\nis still initialising — check GET /status first).",
				"tags": [
					"System"
				],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"basicAuth": []
					}
				],
				"parameters": [
					{
						"description": "Output format: `json` returns a base64 data-URL; `png` returns raw image bytes",
						"in": "query",
						"name": "format",
						"required": false,
						"schema": {
							"default": "json",
							"type": "string",
							"enum": [
								"png",
								"json"
							]
						}
					}
				]
			}
		},
...
		"/chats": {
			"get": {
				"operationId": "GetChats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ChatsResponse"
								}
							}
						}
					},
					"503": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceUnavailableError"
								}
							}
						}
					}
				},
				"description": "List all WhatsApp chats, ordered by most-recent activity (descending timestamp).\n\nPagination uses an offset-based approach — the backend sorts all chats by\ntimestamp descending first, then slices `limit` items starting at `offset`.\n`hasMore` is `true` when there are additional chats beyond the current page.",
				"tags": [
					"Chats"
				],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"basicAuth": []
					}
				],
				"parameters": [
					{
						"description": "Maximum number of chats to return (1–200, default 50)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 50,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Number of chats to skip (0-based, default 0)",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chats/{chatId}/messages": {
			"get": {
				"operationId": "GetChatMessages",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessagesResponse"
								}
							}
						}
					},
					"503": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceUnavailableError"
								}
							}
						}
					}
				},
				"description": "Fetch message history for a specific chat.\n\nUses a custom implementation that bypasses the SDK's getChatById() which\nhas serialization issues with certain chat types (LID-based contacts, etc.).\n\nPagination uses a cursor-based approach — pass the `before` parameter with\nthe Unix-epoch timestamp of the oldest message currently displayed to get\nthe next page of older messages. `hasMore` is `true` when more older messages\nexist beyond the current page.",
				"tags": [
					"Chats"
				],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"basicAuth": []
					}
				],
				"parameters": [
					{
						"description": "WhatsApp chat ID, e.g. `16073041892@c.us` or `123456789@g.us`",
						"in": "path",
						"name": "chatId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Number of messages to fetch (1–200, default 50)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 50,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Optional cursor — only return messages older than this Unix-epoch timestamp",
						"in": "query",
						"name": "before",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chats/{chatId}/read": {
			"post": {
				"operationId": "MarkChatRead",
				"responses": {
					"200": {
						"description": "Chat marked as read",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"ok": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"ok"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"503": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceUnavailableError"
								}
							}
						}
					}
				},
				"description": "Mark all messages in a chat as read.\n\nSends a \"seen\" acknowledgement to WhatsApp for the given chat, clearing\nthe unread badge on the phone and the API.",
				"tags": [
					"Chats"
				],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"basicAuth": []
					}
				],
				"parameters": [
					{
						"description": "WhatsApp chat ID, e.g. `16073041892@c.us` or `123456789@g.us`",
						"in": "path",
						"name": "chatId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "https://[object Object]/"
		}
	],
...
}